cs410videosearchengine/requirements.txt
django
gunicorn
psycopg2
djangorestframework
elasticsearch
google-cloud-speech
google-cloud-storage
sklearn
spacy
nltk
django-extensions
werkzeug
grpcio
protobuf
redis
celery
skipthoughts
summa
boto3
